update to use startx ~/.xinitrc <WM> after updating to gentoo-sources kernel 3.12.13 to 3.14.14 and xorg-server to 1.14.3-r2 3.15.0
The previous cdmrc doesn't work with 0.5.3 and 0.6.1_pre20130419. It might not be the kernel, but when I booted back from 3.12.13, without changing anything, cdm can start X without an issue.

It might not need to use 0.6.1, but after being unable to start X, updating cdm seemed like a logic choice, although it's not the problem source in the end.